fix acceptance tests in local environment

+++ tests/acceptance/environment.py
@@ -1,29 +1,52 @@
 import os
 import time
-
+import sqlalchemy
+import redis
 from celery import Celery
-import deeptracy_core.dal.database as database
+from urllib.parse import urlparse
 
 
 def before_all(context):
-    os.system('docker-compose -f tests/acceptance/docker-compose.yml rm -f')
-    os.system('docker-compose -f tests/acceptance/docker-compose.yml -p deeptracy_acceptance up -d --build')
-    time.sleep(5)
+    """For LOCAL_BEHAVE you need to setup the environment manually"""
 
-    # set environment
-    database.DATABASE_URI = "postgresql://postgres:postgres@127.0.0.1:5432/deeptracy_test"
-    context.BROKER_URI = "redis://127.0.0.1:6379/1"
-    context.SCAN_PATH = "/tmp/deeptracy_test"
+    if os.environ.get('LOCAL_BEHAVE', None) is None:
+        # set environment
+        os.environ['DATABASE_URI'] = 'postgresql://postgres:postgres@127.0.0.1:5432/deeptracy_test'
+        os.environ['BROKER_URI'] = 'redis://127.0.0.1:6379'
+        os.environ['SCAN_PATH'] = '/tmp/deeptracy'
+
+        os.system('docker-compose -f tests/acceptance/docker-compose.yml rm -f')
+        os.system('docker-compose -f tests/acceptance/docker-compose.yml up -d --build')
+        time.sleep(5)
+
+    context.BROKER_URI = os.environ['BROKER_URI']
+    context.SCAN_PATH = os.environ['SCAN_PATH']
 
     if not os.path.exists(context.SCAN_PATH):
         os.makedirs(context.SCAN_PATH)
 
-    database.db.init_engine()  # Init database engine
-
-    context.celery = Celery('deeptracy_test', broker=context.BROKER_URI)
+    context.engine = sqlalchemy.create_engine(os.environ['DATABASE_URI'])
+    context.redis_db = _setup_redis(context.BROKER_URI)
+    context.celery = Celery('deeptracy', broker=context.BROKER_URI)
 
 
 def after_all(context):
-    os.system('docker-compose -f tests/acceptance/docker-compose.yml kill')
-    os.system('docker-compose -f tests/acceptance/docker-compose.yml rm -f')
-    os.system('rm -rf {}'.format(context.SCAN_PATH))
+    if os.environ.get('LOCAL_BEHAVE', None) is None:
+        os.system('docker-compose -f tests/acceptance/docker-compose.yml kill')
+        os.system('docker-compose -f tests/acceptance/docker-compose.yml rm -f')
+        os.system('rm -rf {}'.format(context.SCAN_PATH))
+
+
+def _setup_redis(uri: dict) -> redis.StrictRedis:
+    _, netloc, path, _, _, _ = urlparse(uri)
+
+    host, port = netloc.split(":", maxsplit=1)
+
+    if not path:
+        path = 0
+
+    redis_db = redis.StrictRedis(host=host,
+                                 port=port,
+                                 db=path)
+
+    return redis_db
